feat: refactor chat UI components into separate widget files
