+++ test/mocks.mocks.dart
@@ -401,6 +401,7 @@ class MockOpenOrdersRepository extends _i1.Mock
 /// A class which mocks [SharedPreferencesAsync].
 ///
 /// See the documentation for Mockito's code generation for more information.
+// ignore: must_be_immutable
 class MockSharedPreferencesAsync extends _i1.Mock
     implements _i11.SharedPreferencesAsync {
   MockSharedPreferencesAsync() {
